requirements.txt
Flask==3.0.0
flask-cors==4.0.0
# LangGraph and LangChain for AI agent
langgraph==0.2.28
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.10
# Playwright for browser automation (replaces Selenium)
playwright==1.41.0
# Streamlit for deployment
streamlit==1.29.0
# Additional dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
fpdf2==2.7.9

